+++ physx/src/triangle_mesh.rs
@@ -14,9 +14,7 @@ impl TriangleMesh {
     /// Owner's own the pointer they wrap, using the pointer after dropping the Owner,
     /// or creating multiple Owners from the same pointer will cause UB.  Use `into_ptr` to
     /// retrieve the pointer and consume the Owner without dropping the pointee.
-    pub(crate) unsafe fn from_raw(
-        ptr: *mut physx_sys::PxTriangleMesh,
-    ) -> Option<Owner<TriangleMesh>> {
+    pub unsafe fn from_raw(ptr: *mut physx_sys::PxTriangleMesh) -> Option<Owner<TriangleMesh>> {
         Owner::from_raw(ptr as *mut Self)
     }
 }